cleanup and generalise config methods
